feat: code refactor
Refactored the code to split out the MB header generation from the MB
Info reading. The MB Header is used to inform the bootloader of
information that the OS desires on boot, while the MB Info data is
provided by the bootloader to the OS. Nothing between the MB Header
Struct and the MB Info struct are shared, and so have been moved to
separate submodules.